Separate consoles to different sources
